+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Builder properly parses CVSS version instead of hardcoding `cvss_v3`
+
 ## [0.1.0b4] - 2026-01-10
 
 ### Changed